Add support for Python3
There is still a segmentation fault when saving files with a
non-existing name. I guess this is a problem that the old file is still
open but it is renamed to the new file. The segmentation fault does not
occur during the callbacks. However, the files are written, so Symoro
is basically working under Python3.